refactor: relocate tests to Go's idiomatic layout, colocated with the packages they cover
Every test lived under tests/unit/ and tests/integration/ as two flat,
unrelated external package trees (package unit_test, package integration),
rather than beside the source files they exercise. That had real costs, not
just stylistic ones:

- Several tests were permanently unreachable and existed only as t.Skip
  stubs, some with the real test logic commented out waiting for a home
  that never came: setupLogging (cmd/ntfy-to-slack), getEnvOrDefault/
  getEnvIntOrDefault (config), and processMessage/handleMessageEvent/
  createDefaultMessage (processor) are all unexported, so an external test
  package could never call them directly.
- Coverage required -coverpkg=./... on every invocation (Makefile and CI)
  to attribute coverage back to internal/* at all, since the tests weren't
  in the packages being measured -- one more thing to keep in sync, and it
  had already drifted once (make test-coverage vs. CI, fixed earlier this
  session).
- MockConfigProvider was defined twice, once in tests/unit and once in
  tests/integration, requiring synchronized edits on every config.Provider
  change (this happened twice this session, for the auth and shutdown
  changes).

Moved every test file to live in the same directory as its source package,
using external test packages (e.g. `package config_test` in
internal/config/) except where a file specifically needs unexported access,
which now uses an internal package instead (internal/config/internal_test.go
is `package config`, for instance). Files that mixed cases from multiple
packages in one file (tests/unit/error_test.go,
tests/unit/error_messaging_test.go) because they lived in a shared flat
tree are split into per-package files.

Added internal/testutil (a regular, non-_test.go package) holding the mocks
previously duplicated or redefined per file -- MockHTTPClient,
MockConfigProvider, MockMessageSender, MockPostProcessor, CustomErrorSender
-- so every package's tests import one shared definition. Dropped
MockNtfyClient and MockStreamProcessor rather than carry them into the new
package: grep confirmed neither was ever actually used anywhere in the
suite.

Un-skipped every test that was blocked purely by package visibility, with
real implementations instead of the old skip stubs:
- TestSetupLogging (cmd/ntfy-to-slack/main_test.go) now actually sets a
  level and reads it back via slog.SetLogLoggerLevel's return value, and
  checks the invalid-LOG_LEVEL warning text.
- TestGetEnvOrDefault/TestGetEnvIntOrDefault (internal/config/internal_test.go)
  call the functions directly.
- TestMessageProcessor_processMessage, _handleMessageEvent, and
  _createDefaultMessage (internal/processor/processor_test.go) call those
  methods directly instead of only being reachable through ProcessStream.

Updated the Makefile (dropped test-unit/test-integration, which were purely
a tests/ directory distinction with nothing under them to separate anymore;
dropped -coverpkg=./..., no longer needed once tests are colocated),
.github/workflows/test.yml and release.yml, README.md, CLAUDE.md, and
RELEASES.md for the new `go test ./...` invocation and layout.

Verified with `go build ./...`, `go vet ./...`, and `go test -race ./...`
after every package's move (not just at the end), plus a full final pass;
every package passes, including the previously-dead-skip tests now running
for real. Coverage without -coverpkg is 80.9% overall, same order as
before relocation and comfortably above CI's 70% gate.
