+++ .github/workflows/test.yml
@@ -49,7 +49,7 @@ jobs:
         run: go vet ./...
 
       - name: Run tests
-        run: go test -v -race -coverprofile=coverage.out -coverpkg=./... ./tests/...
+        run: go test -v -race -coverprofile=coverage.out ./...
 
       - name: Coverage check
         run: |
