+++ internal/app/app_test.go
@@ -1,0 +1,73 @@
+package app_test
+
+import (
+	"bytes"
+	"io"
+	"os"
+	"strings"
+	"testing"
+
+	"github.com/ozskywalker/ntfy-to-slack/internal/app"
+	"github.com/ozskywalker/ntfy-to-slack/internal/testutil"
+)
+
+func TestNewApp(t *testing.T) {
+	config := &testutil.MockConfigProvider{
+		Domain:     "test.ntfy.sh",
+		Topic:      "test-topic",
+		Auth:       "auth-token",
+		WebhookURL: "https://hooks.slack.com/test",
+	}
+
+	appInstance := app.New(config, "test-version")
+
+	if appInstance == nil {
+		t.Error("app.New() returned nil")
+	}
+}
+
+// TestPrintHelp_Output captures PrintHelp's stdout and checks the Flags
+// section (generated by config.FlagUsage()) and the surrounding sections
+// are all present. setupLogging (the other historically-untestable piece
+// of cmd/ntfy-to-slack) has its own coverage in
+// cmd/ntfy-to-slack/main_test.go, now that it's colocated with package main
+// and reachable directly.
+func TestPrintHelp_Output(t *testing.T) {
+	config := &testutil.MockConfigProvider{}
+	appInstance := app.New(config, "test-version")
+
+	oldStdout := os.Stdout
+	r, w, _ := os.Pipe()
+	os.Stdout = w
+
+	appInstance.PrintHelp()
+
+	w.Close()
+	os.Stdout = oldStdout
+
+	var buf bytes.Buffer
+	io.Copy(&buf, r)
+	output := buf.String()
+
+	// The Flags section is generated by config.FlagUsage() from the actual
+	// flag.FlagSet (see config.buildFlagSet), which renders flag names with
+	// a single leading dash (Go's flag package accepts both "-x" and "--x"
+	// when parsing, but only ever prints the single-dash form itself).
+	expectedStrings := []string{
+		"ntfy-to-slack",
+		"Forwards ntfy.sh messages to Slack",
+		"Usage:",
+		"-ntfy-domain",
+		"-ntfy-topic",
+		"-slack-webhook",
+		"Environment Variables:",
+		"NTFY_DOMAIN",
+		"SLACK_WEBHOOK_URL",
+	}
+
+	for _, expected := range expectedStrings {
+		if !strings.Contains(output, expected) {
+			t.Errorf("Expected help output to contain %q, but it didn't. Output: %s", expected, output)
+		}
+	}
+}
