+++ internal/app/idlewatchdog_test.go
@@ -1,4 +1,4 @@
-package unit_test
+package app_test
 
 import (
 	"errors"
