+++ internal/config/config_postprocessor_test.go
@@ -1,4 +1,4 @@
-package unit_test
+package config_test
 
 import (
 	"testing"
