+++ internal/config/edge_cases_test.go
@@ -1,14 +1,16 @@
-package unit_test
+package config_test
+
+// These cases originally lived in cmd/ntfy-to-slack's test file under the
+// theory that they exercised "what main does", but they only ever call
+// config.New/Validate directly ("Test configuration parsing directly since
+// we can't easily test main() exit codes" -- a comment from the original),
+// so they're config tests in every way that matters and belong here.
 
 import (
-	"bytes"
-	"fmt"
-	"io"
 	"os"
 	"strings"
 	"testing"
 
-	"github.com/ozskywalker/ntfy-to-slack/internal/app"
 	"github.com/ozskywalker/ntfy-to-slack/internal/config"
 )
 
@@ -148,73 +150,6 @@ func TestMain_EnvironmentPrecedence(t *testing.T) {
 
 	if cfg.GetSlackWebhookURL() != "https://hooks.slack.com/env" {
 		t.Errorf("Expected env var to be used when flag not provided, got: %s", cfg.GetSlackWebhookURL())
-	}
-}
-
-func TestSetupLogging_EdgeCases(t *testing.T) {
-	tests := []string{
-		"DEBUG",   // Different case
-		"Info",    // Mixed case
-		"WARN",    // Upper case
-		"ERROR",   // Upper case
-		"trace",   // Invalid level
-		"verbose", // Invalid level
-		"",        // Empty string
-		"   ",     // Whitespace
-	}
-
-	for _, logLevel := range tests {
-		t.Run(fmt.Sprintf("level_%s", logLevel), func(t *testing.T) {
-			// setupLogging is not exported, skip this test
-			t.Skip("setupLogging is not exported")
-		})
-	}
-}
-
-func TestPrintHelp_Output(t *testing.T) {
-	// Capture stdout
-	oldStdout := os.Stdout
-	r, w, _ := os.Pipe()
-	os.Stdout = w
-
-	// printHelp is not exported, use app.PrintHelp instead
-	cfg, err := config.New([]string{"--ntfy-topic", "test", "--slack-webhook", "https://test.com"})
-	if err != nil {
-		t.Fatalf("Failed to create config: %v", err)
-	}
-	appInstance := app.New(cfg, "test-version")
-	appInstance.PrintHelp()
-
-	// Restore stdout
-	w.Close()
-	os.Stdout = oldStdout
-
-	// Read captured output
-	var buf bytes.Buffer
-	io.Copy(&buf, r)
-	output := buf.String()
-
-	// Verify output contains expected information. The Flags section is
-	// generated by config.FlagUsage() from the actual flag.FlagSet (see
-	// config.buildFlagSet), which renders flag names with a single leading
-	// dash (Go's flag package accepts both "-x" and "--x" when parsing, but
-	// only ever prints the single-dash form itself).
-	expectedStrings := []string{
-		"ntfy-to-slack",
-		"Forwards ntfy.sh messages to Slack",
-		"Usage:",
-		"-ntfy-domain",
-		"-ntfy-topic",
-		"-slack-webhook",
-		"Environment Variables:",
-		"NTFY_DOMAIN",
-		"SLACK_WEBHOOK_URL",
-	}
-
-	for _, expected := range expectedStrings {
-		if !strings.Contains(output, expected) {
-			t.Errorf("Expected help output to contain %q, but it didn't. Output: %s", expected, output)
-		}
 	}
 }
