+++ internal/config/postprocessor_integration_test.go
@@ -1,4 +1,4 @@
-package integration
+package config_test
 
 import (
 	"encoding/json"
