--- tests/integration/integration_test.go
@@ -1,4 +1,0 @@
-package integration
-
-// This file is kept for potential future integration tests
-// Legacy tests have been removed and replaced with proper dependency injection tests
