--- tests/unit/processor_postprocessor_test.go
@@ -1,186 +1,0 @@
-package unit_test
-
-import (
-	"errors"
-	"strings"
-	"testing"
-
-	"github.com/ozskywalker/ntfy-to-slack/internal/config"
-	"github.com/ozskywalker/ntfy-to-slack/internal/processor"
-)
-
-func TestNewMessageProcessorWithPostProcessor(t *testing.T) {
-	sender := &MockMessageSender{}
-	postProcessor := &MockPostProcessor{}
-
-	processor := processor.NewWithPostProcessor(sender, postProcessor)
-
-	if processor == nil {
-		t.Error("Expected processor but got nil")
-	}
-
-	// Note: internal fields are not exported, so we can't test them directly
-	// The fact that NewWithPostProcessor() doesn't panic and returns a non-nil processor is sufficient
-}
-
-func TestMessageProcessor_ProcessWithPostProcessor(t *testing.T) {
-	tests := []struct {
-		name                string
-		postProcessorFunc   func(message *config.NtfyMessage) (*config.SlackMessage, error)
-		inputMessage        *config.NtfyMessage
-		expectedSlackText   string
-		postProcessorCalled bool
-		shouldUseDefault    bool
-	}{
-		{
-			name: "successful post-processing",
-			postProcessorFunc: func(message *config.NtfyMessage) (*config.SlackMessage, error) {
-				return &config.SlackMessage{Text: "Custom: " + message.Title + " - " + message.Message}, nil
-			},
-			inputMessage: &config.NtfyMessage{
-				Event:   "message",
-				Title:   "Alert",
-				Message: "System down",
-			},
-			expectedSlackText:   "Custom: Alert - System down",
-			postProcessorCalled: true,
-		},
-		{
-			name: "post-processor error falls back to default",
-			postProcessorFunc: func(message *config.NtfyMessage) (*config.SlackMessage, error) {
-				return nil, errors.New("processing failed")
-			},
-			inputMessage: &config.NtfyMessage{
-				Event:   "message",
-				Title:   "Alert",
-				Message: "System down",
-			},
-			expectedSlackText:   "*Alert*: System down",
-			postProcessorCalled: true,
-			shouldUseDefault:    true,
-		},
-		{
-			name: "message without title",
-			postProcessorFunc: func(message *config.NtfyMessage) (*config.SlackMessage, error) {
-				return &config.SlackMessage{Text: "Processed: " + message.Message}, nil
-			},
-			inputMessage: &config.NtfyMessage{
-				Event:   "message",
-				Title:   "",
-				Message: "Simple message",
-			},
-			expectedSlackText:   "Processed: Simple message",
-			postProcessorCalled: true,
-		},
-	}
-
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			// handleMessageEvent is not exported, skip this test as it tests internal implementation
-			t.Skip("handleMessageEvent is not exported - test internal implementation")
-		})
-	}
-}
-
-func TestMessageProcessor_ProcessStream_WithPostProcessor(t *testing.T) {
-	input := `{"event":"open","topic":"test"}
-{"id":"msg1","time":1640995200,"event":"message","topic":"test","title":"Alert","message":"System issue"}
-{"event":"keepalive","topic":"test"}
-{"id":"msg2","time":1640995201,"event":"message","topic":"test","title":"","message":"Simple message"}`
-
-	sender := &MockMessageSender{}
-	postProcessor := &MockPostProcessor{
-		ProcessFunc: func(message *config.NtfyMessage) (*config.SlackMessage, error) {
-			return &config.SlackMessage{Text: "🔔 " + message.Title + ": " + message.Message}, nil
-		},
-	}
-
-	processor := processor.NewWithPostProcessor(sender, postProcessor)
-	reader := strings.NewReader(input)
-
-	err := processor.ProcessStream(reader)
-	if err != nil {
-		t.Errorf("Unexpected error: %v", err)
-	}
-
-	// Should have processed 2 message events
-	if postProcessor.CallCount != 2 {
-		t.Errorf("Expected post-processor to be called 2 times, got %d", postProcessor.CallCount)
-	}
-
-	if len(sender.SentMessages) != 2 {
-		t.Errorf("Expected 2 sent messages, got %d", len(sender.SentMessages))
-		return
-	}
-
-	expectedMessages := []string{
-		"🔔 Alert: System issue",
-		"🔔 : Simple message",
-	}
-
-	for i, expected := range expectedMessages {
-		if sender.SentMessages[i].Text != expected {
-			t.Errorf("Message %d: expected %q, got %q", i, expected, sender.SentMessages[i].Text)
-		}
-	}
-}
-
-func TestMessageProcessor_WithoutPostProcessor(t *testing.T) {
-	// handleMessageEvent is not exported, skip this test
-	t.Skip("handleMessageEvent is not exported - test internal implementation")
-}
-
-func TestMessageProcessor_createDefaultMessage(t *testing.T) {
-	// createDefaultMessage is not exported, skip this test
-	t.Skip("createDefaultMessage is not exported - test internal implementation")
-
-	/*
-		tests := []struct {
-			name         string
-			message      *config.NtfyMessage
-			expectedText string
-		}{
-			{
-				name: "message with title",
-				message: &config.NtfyMessage{
-					Title:   "Alert",
-					Message: "System issue",
-				},
-				expectedText: "*Alert*: System issue",
-			},
-			{
-				name: "message without title",
-				message: &config.NtfyMessage{
-					Title:   "",
-					Message: "Simple message",
-				},
-				expectedText: "Simple message",
-			},
-			{
-				name: "message with empty title",
-				message: &config.NtfyMessage{
-					Title:   "",
-					Message: "Another message",
-				},
-				expectedText: "Another message",
-			},
-		}
-
-		processor := processor.New(&MockMessageSender{})
-
-		for _, tt := range tests {
-			t.Run(tt.name, func(t *testing.T) {
-				result := processor.createDefaultMessage(tt.message)
-
-				if result == nil {
-					t.Error("Expected result but got nil")
-					return
-				}
-
-				if result.Text != tt.expectedText {
-					t.Errorf("Expected text %q, got %q", tt.expectedText, result.Text)
-				}
-			})
-		}
-	*/
-}
